wip:migrate script folder to cli commands and refactor cli file structure
