+++ CLI/hotReload.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"fmt"
@@ -9,30 +9,38 @@ import (
 	"sync"
 )
 
+type HotReloadCommand struct {
+	cli *GothicCli
+}
+
 var (
 	templCmd *exec.Cmd
 	airCmd   *exec.Cmd
 	mu       sync.Mutex
 )
 
-func main() {
+func NewHotReloadCommandCli() HotReloadCommand {
+	return HotReloadCommand{}
+}
+
+func (command *HotReloadCommand) HotReload() {
 	// Start the commands and wait for their completion
 	var wg sync.WaitGroup
 	wg.Add(2)
 
 	go func() {
 		defer wg.Done()
-		runCommand("templ", []string{"generate", "--watch", "--proxy=http://localhost:8080"})
+		command.runCommand("templ", []string{"generate", "--watch", "--proxy=http://localhost:8080"})
 	}()
-
+	// TODO remove air and replace for custom go script
 	go func() {
 		defer wg.Done()
-		runCommand("air", []string{})
+		command.runCommand("air", []string{})
 	}()
 	select {}
 }
 
-func runCommand(name string, args []string) {
+func (command *HotReloadCommand) runCommand(name string, args []string) {
 	cmd := exec.Command(name, args...)
 	mu.Lock()
 	if name == "templ" {
@@ -45,19 +53,19 @@ func runCommand(name string, args []string) {
 	// Create pipes for stdout and stderr
 	stdout, err := cmd.StdoutPipe()
 	if err != nil {
-		handleError(fmt.Sprintf("Error creating stdout pipe for command %s", name), err)
+		command.handleError(fmt.Sprintf("Error creating stdout pipe for command %s", name), err)
 		return
 	}
 
 	stderr, err := cmd.StderrPipe()
 	if err != nil {
-		handleError(fmt.Sprintf("Error creating stderr pipe for command %s", name), err)
+		command.handleError(fmt.Sprintf("Error creating stderr pipe for command %s", name), err)
 		return
 	}
 
 	// Start the command
 	if err := cmd.Start(); err != nil {
-		handleError(fmt.Sprintf("Error starting command %s", name), err)
+		command.handleError(fmt.Sprintf("Error starting command %s", name), err)
 		return
 	}
 
@@ -69,16 +77,16 @@ func runCommand(name string, args []string) {
 
 	// Wait for the command to finish
 	if err := cmd.Wait(); err != nil {
-		handleError(fmt.Sprintf("Command %s finished with error", name), err)
+		command.handleError(fmt.Sprintf("Command %s finished with error", name), err)
 	}
 }
 
-func handleError(message string, err error) {
+func (command *HotReloadCommand) handleError(message string, err error) {
 	log.Println(message, err)
-	gracefulShutdown()
+	command.gracefulShutdown()
 }
 
-func gracefulShutdown() {
+func (command *HotReloadCommand) gracefulShutdown() {
 	mu.Lock()
 	defer mu.Unlock()
