+++ CLI/templates.go
@@ -1,4 +1,4 @@
-package templates
+package cli
 
 import (
 	"embed"
