+++ Templates/template-custom-domain-with-arn.yaml
