+++ go.mod
@@ -9,4 +9,7 @@ require (
 	github.com/teris-io/shortid v0.0.0-20220617161101-71ec9f2aa569
 )
 
-require golang.org/x/image v0.24.0 // indirect
+require (
+	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646 // indirect
+	golang.org/x/image v0.26.0 // indirect
+)
